% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/An_t_1.R
\name{An_t_1}
\alias{An_t_1}
\title{Dynamic System at An(t=1)}
\usage{
An_t_1(Ao, steps, .fun)
}
\arguments{
\item{Ao}{number, f(x) value, or A(n) value, of the initial condition at time 0.}

\item{steps}{integer, the number of times steps into the future.}

\item{.fun}{function, a mathematical function defined by function().}
}
\value{
A dataframe with time steps (t), Ao recursive value of An(t), and
An_t_1 the value based on the lagged An(t), previous result.
}
\description{
Derive the limit at point a.
}
\details{
Builds a recursive solution with A(t=1) = A(0), or the initial condition.
Can be used to test the fixed points of a system by building the cobweb from this
data.
}
\note{
No imports
}
\examples{
{
 fx2 <- function(x) {x^2}
 An1 <- An_t_1(Ao = 0, steps = 100, .fun = fx2)
}

}
