% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck_of_cards.R
\name{deck_of_cards}
\alias{deck_of_cards}
\alias{shuffle_cards}
\alias{full_house_prob}
\alias{two_of_kind}
\title{Deck of Cards}
\usage{
deck_of_cards()

shuffle_cards(deck, hand_size, ...)

full_house_prob(sims, ...)

two_of_kind(sims, ...)
}
\arguments{
\item{deck}{dataframe, dataframe with a deck of cards.  Can be built with
\code{deck_of_cards()}.}

\item{hand_size}{number, number of cards for each hand or run.}

\item{...}{arguments passed to \code{shuffle_cards()}, which can also pass args to
\code{dplyr::sample_n()}.}

\item{sims}{number, the number of simulations to run in for loop.}
}
\value{
A vector of points that replicate the limit as they approach a.
}
\description{
Functions related to generating deck of cards data and probabilities
for different spread of hands.
}
\details{
This is a test.

This is test2
}
\examples{
{
 # Instantiate a deck
 deck_cards <- deck_of_cards()
 # Shuffle Cards, select hand size
 dc_shuffle <- shuffle_cards(deck = deck_cards, hand_size = 5)
 # Generate probability for Full House
 fh_prob <- full_house_prob(sims = 10000, deck = deck_cards, hand_size = 5)
 # Genearte probability for Two of a Kind
 two_of_kind(sims = 50, deck = NA, hand_size = 5)
}
}
