% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dy_dx.R
\name{dy_dx}
\alias{dy_dx}
\title{Derivative of f(x)}
\usage{
dy_dx(.fun, ...)
}
\arguments{
\item{.fun}{function, a mathematical function defined by function()}

\item{...}{paramaters passed on to the \code{limit_at_a()} function.  The needed
parameters are:
\itemize{
\item \code{a} number, numerical vector of X's
\item `h`` number, is the range around an X's used to derive the slope or limit.
\item \code{step_size} number, the step size in deriving the slope or limit.
}}
}
\value{
Creates a vector of output for dy/dx of function f(x).  Combined with
}
\description{
Generate the data derivative of f(x)
}
\details{
Given a vector of X's and a function, this function will generate
the Y's for dy/dx, or the output of the derivative.
}
\examples{
{
  fx2 <- function(x) {x^6} # The function
  x_vector <- seq(-100,100, by = 10) # The X vector
  dy_dx(.fun = fx2, a = x_vector, h = 1, step_size = 0.5) # The Derivative
}

}
