% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_a_h.R
\name{f_a_h}
\alias{f_a_h}
\title{Centered Vector}
\usage{
f_a_h(a, h, step_size, ...)
}
\arguments{
\item{a}{number, that centers the vector}

\item{h}{number, the point from the center will be is +/- from the center.}

\item{step_size}{number, size of steps between numbers.}

\item{...}{arguments passed to \code{seq()} function.}
}
\value{
A vector from a-h to a+h, centered on a, by step size \code{step_size.}
}
\description{
Create vector centered on \code{a} and bounded by \code{h}, by step size \code{step_size}.
}
\details{
Creates a vector from a-h to a+h, centered on a, by step size \code{step_size}.
}
\examples{
{
 vector <- f_a_h(a = 0, h = 10, step_size = 1)
}

}
