% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_at_a.R
\name{limit_at_a}
\alias{limit_at_a}
\title{The Limit at a}
\usage{
limit_at_a(a, h, .fun, step_size, ...)
}
\arguments{
\item{a}{number, the number which is being limited.}

\item{h}{number, the point from the center, or a, that is +/- from the center.}

\item{.fun}{function, a mathematical function defined by function().}

\item{step_size}{number, size of steps between numbers, or a and h.}

\item{...}{arguments passed to \code{seq()}}
}
\value{
A vector of points that replicate the limit as they approach a.
}
\description{
Derive the limit at point a.
}
\details{
Combines the \code{f_a_h()} and \code{limit_dx1_dx2()} to generate a vector
of numbers as the numbers approach a shows the limit to a.
}
\examples{
{
 fx2 <- function(x) {x^2}
 x_vector <- seq(-2*pi,2*pi, by = (pi/8))
 limit_at_a(a = 0, h = 10, step_size = 0.1, .fun = fx2)
}

}
