% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_x_roots.R
\name{quad_x_roots}
\alias{quad_x_roots}
\title{Quadratic Roots}
\usage{
quad_x_roots(a, b, c)
}
\arguments{
\item{a}{number, coefficient of x^2}

\item{b}{number, coefficient of x}

\item{c}{number, constanct C}
}
\value{
A solution points, if exists, for a quadratic function.
}
\description{
Derive the quadratic roots for a quadratic function.
}
\details{
Delivers roots conditioned on the sign of the discriminant =
(b^2) - (4\emph{a}c)
}
\examples{
{
 roots <- quad_x_roots(1,7,5)
}

}
