% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_x1_x2.R
\name{slope_x1_x2}
\alias{slope_x1_x2}
\title{The Slope (Limit) at a Point}
\usage{
slope_x1_x2(x1, x2, .fun)
}
\arguments{
\item{x1}{numeric, limit number}

\item{x2}{numeric, the number near the x1}

\item{.fun}{function, a function to find the limit, or slope}
}
\value{
A number, or limit of x1 at x2.
}
\description{
Find the slope, rise over run, at a point x1 from another point
x2 for a function.
}
\details{
Find the slope, or rise over run, at a point x1 from another point
x2 for a function.
}
\examples{
{
 fx2 <- function(x) {x^2}
 slope_x1_x2(x1 = 1, x2 = 4, .fun = fx2)
}

}
